Fix compile errors with some compilers.

+++ src/modules/font/GenericShaper.cpp
@@ -105,7 +105,7 @@ void GenericShaper::computeGlyphPositions(const ColoredCodepoints &codepoints, R
 
 		if (colorToAdd.hasValue && colors && positions)
 		{
-			IndexedColor c = {colorToAdd.value, positions->size()};
+			IndexedColor c = {colorToAdd.value, (int) positions->size()};
 			colors->push_back(c);
 			colorToAdd.clear();
 		}
